+++ dolphinscheduler-common/pom.xml
@@ -587,10 +587,7 @@
 			<scope>compile</scope>
         </dependency>
 
-		<dependency>
-			<groupId>org.springframework</groupId>
-			<artifactId>spring-context</artifactId>
-		</dependency>
+
 
 		<dependency>
 			<groupId>org.codehaus.janino</groupId>
