+++ dolphinscheduler-service/src/main/java/org/apache/dolphinscheduler/service/utils/Preconditions.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.dolphinscheduler.common.utils;
+package org.apache.dolphinscheduler.service.utils;
 
 import org.springframework.lang.Nullable;
